Add Config resource definition and interfaces for various resource types
- Introduced a new resource type "Config" in Resourcedefs with appropriate TypeMeta and versions.
- Added interfaces for multiple resource types including Application, AppProject, BackupJob, BackupRun, Certificate, ClusterComplianceReport, ClusterVulnerabilityReport, ConfigAuditReport, DaemonSet, Datacenter, Deployment, Endpoints, ExposedSecretReport, Ingress, IngressClass, Node, PersistentVolumeClaim, Pod, PolicyReport, Provider, RbacAssessmentReport, ReplicaSet, Route, SbomReport, Service, SlackMessage, StatefulSet, StorageClass, TanzuKubernetesCluster, TanzuKubernetesRelease, VirtualMachineClass, KubernetesCluster, KubernetesMachineClass, Machine, Unknown, and their respective resource definitions.
