+++ pkg/rorresources/rortypes/resourcedef_clustervulnerabilityreport.go
@@ -100,3 +100,8 @@ func (s VulnerabilityDismissalReason) TSName() string {
 func (r *ResourceClusterVulnerabilityReport) Get() *ResourceClusterVulnerabilityReport {
 	return r
 }
+
+// ClusterVulnerabilityReportinterface represents the interface for resources of the type clustervulnerabilityreport
+type ClusterVulnerabilityReportinterface interface {
+	Get() *ResourceClusterVulnerabilityReport
+}
