+++ pkg/rorresources/rortypes/resourcedef_vulnerabilityevent.go
@@ -17,3 +17,8 @@ type ResourceVulnerabilityEventSpec struct {
 func (r *ResourceVulnerabilityEvent) Get() *ResourceVulnerabilityEvent {
 	return r
 }
+
+// VulnerabilityEventinterface represents the interface for resources of the type vulnerabilityevent
+type VulnerabilityEventinterface interface {
+	Get() *ResourceVulnerabilityEvent
+}
