+++ pkg/rorresources/rortypes/resourcedef_vulnerabilityreport.go
@@ -44,3 +44,9 @@ func (r *ResourceVulnerabilityReport) ApplyInputFilter(cr *CommonResource) error
 func (r *ResourceVulnerabilityReport) Get() *ResourceVulnerabilityReport {
 	return r
 }
+
+// VulnerabilityReportinterface represents the interface for resources of the type vulnerabilityreport
+type VulnerabilityReportinterface interface {
+	ApplyInputFilter(cr *CommonResource) error
+	Get() *ResourceVulnerabilityReport
+}
